Large revision. See commit details for more
This commit brings QRatittude to a new world order.
express dependency is shot in favor of restify,
and application entry point is now organized
using dependency injection.

Server implementation is now in server.js,
and all logic is tied together in app.js,
which still serves as the entry point.

The server is not fully functional, but
thankfully it is much easier to extend
and maintain. Note the following:

1. The restify server does not have an all() method, so those lines
   were commented in server.js for the time being.

2. On that note, everything in express went out the window.
   Calls to use() have been removed.

3. curl-test.sh fires a request ok, but the asset to be inserted
   is said to be undefined.